Modify names of variables and routine assoc, with min heap of chunks
In future changes we will use a min heap of processes (with weights
defined by estimated cost of chunks assigned to a process). Here the
min heap of chunks used when assigning columns to chunks is renamed to
cheap. Names of associated variables and routines are likewise
modified, as is the corresponding documentation.

[BFB]