Improve load balancing when chunks have different costs
The assign_chunks algorithm is modified to balance the total estimated
cost per process while retaining the current approach of attempting to
minimizing the communication costs in dp_coupling. This is implemented
by introducing a min heap of processes ranked by the estimated cost
based on the estimated cost of the chunks assigned to each
process. Chunks are assigned in order of decreasing estimated cost,
and estimated process cost is adjusted when assigning a chunk to
thread 0 (assuming a wrap map of chunks to threads within a process
and perfectly parallel computation between threads). If there are
multiple candidate processes to which a chunk can be assigned, the
chunk is assigned to a process which would minimize the
communication cost in d_p_coupling and p_d_coupling.

[BFB]