Minimize number of columns communicated during coupling
The initial implementation of the new algorithm to assign chunks to
processes in a more load balanced way missed some opportunities to
minimize the number of columns that need to be copied during
couppling between the physics and dynamics. Here a change is made to
require approximately the same number of column transfers as with the
old chunk assignment algorithm when the chunks all have the same
cost. Another change is to delay updating the process min heap after
the estimated cost of a process changes (when a chunk is added to
thread 0) until all other threads have had a chunk assigned to
it. Since chunks are assigned in order of decreasing cost, the
estimated cost of the process will not increase again
until this occurs. The delay allows for addiitonal opportunities to
minimize communication overhead by continuing to regard this process
as having lower cost than it actually has.

Finally, the format of the output in the file atm_chunk_costs.txt is
modifed to include the number of columns in each chunk that do not
have to transferred to other processes during coupling.

[BFB]